feat: Finished 'Modules' module
